refactor(agent-detail): extract ResizeHandle component and clean up AgentDetailView
- Move ResizeHandle functionality to a new shared component for better reusability.
- Simplify AgentDetailView by removing inline ResizeHandle implementation.
- Introduce new chatComponents file to house chat-related components and utilities.
- Refactor ChatPanel to utilize shared components and improve code organization.
